发版脚本

+++ release.bat
@@ -1,16 +1,17 @@
 @echo off
-set target=E:\workspace\WebSQL2
+set target=D:\workspace\WebSQL2
 
 tskill WebSql
 
 go build -o WebSql.exe main.go
-del /q %target%\WebSql.exe
-mv WebSql.exe %target%
+if exist %target%\WebSql.exe del /q %target%\WebSql.exe
+move /Y WebSql.exe %target%
 
 cd web-src
 call npm run build
-del /s /q %target%\static\*
-xcopy /Y /e dist %target%\static
+if exist %target%\static del /s /q %target%\static\*
+if not exist %target%\static mkdir %target%\static
+xcopy /Y /E dist\* %target%\static
 
 cd ..
 start /b %target%\WebSql.exe
